+++ src/echoes/of/the/dead/ChooseChar.java
@@ -181,22 +181,22 @@ public class ChooseChar extends javax.swing.JFrame implements MouseInteractable 
             charType = "knight";
             scene.character.setCharacterType(charType);
             scene.character.animator.importSprites("character_asset", "idle", 18, 6);
-            scene.character.setPosX((int)(width * 0.32)); // Update position based on width
-            scene.character.setPosY((int)(height * 0.51)); // Update position based on height
+            scene.character.setPosX(width * 0.32); // Update position based on width
+            scene.character.setPosY(height * 0.51); // Update position based on height
             scene.setCurrentSceneIndex(0);
         } else if (source == btn_priest) {
             charType = "priest";
             scene.character.setCharacterType(charType);
             scene.character.animator.importSprites("character_asset", "idle", 18, 6);
-            scene.character.setPosX((int)(width * 0.349)); // Update position based on width
-            scene.character.setPosY((int)(height * 0.49)); // Update position based on height
+            scene.character.setPosX(width * 0.349); // Update position based on width
+            scene.character.setPosY(height * 0.49); // Update position based on height
             scene.setCurrentSceneIndex(0);
         } else if (source == btn_wizard) {
             charType = "wizard";
             scene.character.setCharacterType(charType);
              scene.character.animator.importSprites("character_asset", "idle", 18, 6);
-            scene.character.setPosX((int)(width * 0.34)); // Update position based on width
-            scene.character.setPosY((int)(height * 0.51)); // Update position based on height
+            scene.character.setPosX(width * 0.34); // Update position based on width
+            scene.character.setPosY(height * 0.51); // Update position based on height
             scene.setCurrentSceneIndex(0);
         }else if(source == btn_ok){
             if((nameField.getText().trim().isEmpty())){ // a condition that sends a warning message to the user if they clicked ok when they didn't enter a name --jian
